Applied feedback from @iliqportals
