+++ index.js
@@ -14,7 +14,7 @@ const attachToGulp = require('./lib/attachToGulp');
  * @static
  * @see {@link #module_lib/portal|lib/portal}
  */
-const Portal = require('./lib/portal');
+const Portal = require('./lib/deploy');
 
 module.exports = {
 	attachToGulp,
