+++ lib/attachToGulp.js
@@ -1,5 +1,9 @@
 "use strict";
 
+const env = require('./private/env');
+const log = require('@studyportals/node-log');
+const pjson = require('./../package.json');
+
 /**
  * Overwrite gulp tasks from `@prtl/local-deploy-portal-spici`
  *
@@ -16,10 +20,6 @@
  */
 const attachToGulp = (gulp, opts) =>{
 
-	const env = require('./private/env');
-	const log = require('@studyportals/node-log');
-
-	const pjson = require('./../package.json');
 	log.debug(`Using ${pjson.name}@${pjson.version}`);
 
 	const defaults = {
@@ -35,7 +35,7 @@ const attachToGulp = (gulp, opts) =>{
 
 	opts.enableCompression = opts.env === env.PRD || opts.env === env.STG;
 
-	const Portal = new (require('./portal'))({
+	const Portal = new (require('./deploy'))({
 		to: opts.buildDir,
 		env: opts.env
 	});
